revert! z80: re-fix "block of EI does not inhibit interrupt"
this was "fixed" in sync with vanilla, but the same logic doesn't work
correctly in ZESERUse, so the code base is too much out of sync.

Reverting back at this moment, so interrupt signal inhibited by DD/FD/EI
long enough to be missed is still processed by ZESERUse even much later,
but never missed. (it's a bug, but not as bad as not handling interrupts
at all, right?)